+++ compiler/bootstrap/libec/bootstrap/shortcuts.c
@@ -28,6 +28,7 @@ typedef unsigned __int64 uint64;
 #define __ENDIAN_PAD(x) 0
 #endif
 #include <stdint.h>
+#include <sys/types.h>
 extern void *  __ecereNameSpace__ecere__com__eSystem_New(unsigned int size);
 
 extern void *  __ecereNameSpace__ecere__com__eSystem_New0(unsigned int size);
@@ -611,7 +612,7 @@ unsigned int parseTypeError;
 
 extern struct __ecereNameSpace__ecere__com__Instance * fileInput;
 
-extern int strlen(const char * );
+extern size_t strlen(const char * );
 
 extern int declMode;
